simplify LockResponse
